feature/add-useSearchLogic-hook
