+++ packages/client/src/pages/SearchPage/SearchPage.css
@@ -1,0 +1,9 @@
+.loading-message {
+  color: #8ece3e;
+  font-size: 20px;
+}
+
+.error-message {
+  color: #c21a1a;
+  font-size: 20px;
+}
